requirements.txt
alembic==1.6.5
aniso8601==9.0.1
anyio==3.2.0
apache-airflow==2.1.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
bleach==3.3.0
blinker==1.4
cattrs==1.7.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
cloudscraper==1.2.42
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
entrypoints==0.3
Flask==1.1.2
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
flask-marshmallow==0.14.0
Flask-OpenID==1.2.5
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.16
greenlet==1.0.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
more-itertools==8.8.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.20.3
nxviz==0.6.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==20.9
palettable==3.3.0
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pendulum==2.1.2
pickleshare==0.7.5
Pillow==8.2.0
prison==0.1.3
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psutil==5.8.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.0
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pywin32==301
pywinpty==1.1.2
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==10.4.0
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.5.0
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
unicodecsv==0.14.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==1.0.1
WTForms==2.3.3
